Update status series after review
* Reverted back to integer incrementation only
* Moved the table for story attribute3 to the right of attribute2,
making the save button more visible.
* Added error messages for bad series input
* Added hover icon for the add buttons
* Fixed a bug where two similar series where group together

+++ src/main/java/com/sonymobile/backlogtool/AttributeOption.java
@@ -52,7 +52,7 @@ public class AttributeOption {
     private String color;
     private int compareValue;
     private boolean iconEnabled = true;
-    private Double seriesIncrement;
+    private Integer seriesIncrement;
 
     public AttributeOption() {}
 
@@ -155,11 +155,11 @@ public class AttributeOption {
         this.iconEnabled = iconEnabled;
     }
 
-    public Double getSeriesIncrement() {
+    public Integer getSeriesIncrement() {
         return seriesIncrement;
     }
 
-    public void setSeriesIncrement(Double seriesIncrement) {
+    public void setSeriesIncrement(Integer seriesIncrement) {
         this.seriesIncrement = seriesIncrement;
     }
